requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-dotenv
PyMySQL # MySQL driver for SQLAlchemy
pydantic[email]